requirements.txt
anthropic>=0.28.0
fastapi==0.111.0
uvicorn[standard]==0.29.0
openai==1.30.1
scikit-learn==1.4.2
numpy==1.26.4
python-dotenv==1.0.1
python-multipart==0.0.9
pydantic==2.7.1
pypdf==4.2.0
python-docx==1.1.2
openpyxl==3.1.2
python-pptx==0.6.23
